+++ src/WinFormedge/Formedge.Internal.cs
@@ -203,15 +203,7 @@ public abstract partial class Formedge
         UpdateFormText();
     }
 
-    protected void RegisterResourceHander(WebResourceHandler resourceHandler)
-    {
-        WebView.WebResourceManager.RegisterResourceHander(resourceHandler);
-    }
 
-    protected void UnregisterResourceHander(WebResourceHandler resourceHandler)
-    {
-        WebView.WebResourceManager.UnregisterResourceHander(resourceHandler);
-    }
 }